chore(odyssey-react-mui): update Typography type override location
